Update README.md
# Semantic Segmentation

The aim of the work is to optimize segmentation in histopathology in order to support the doctor in the analysis of intestinal cells and in decision-making using AI-based solutions. Subject of the presented work are studies of segmentation in histopathology using the example of the data sets WARWICK-QU and CRAG-v2 using two U-Net variants. One focus is the development of an experimental project that allows simple parameterization and documentation of results using a configuration file. The resolution of the image-mask pairs, the learning rate, the number of training epochs, the number of blocks of the U-Net and the input filter for the U-Net are verified. Possibilities of post-processing predicted masks are also considered. This allows an image improvement and, above all, an indexing of the glands. The results of a large number of experiments carried out were documented in tabular form and in the form of diagrams. General key figures were selected for the objective evaluation of the forecast results: the Dice index, the F1 score and the weighted shape. The pictorial representation of both the input image-mask-pairs and the predicted masks illustrates the results.
The own, developed project delivers good to very good segmentation results for both data sets WARWICK-QU and CRAG-v2. The using this concept for similar segmentation tasks is recommended.
Keywords: Histopathology, semantic segmentation, neural networks, U-Net, Dataset, WARWICK-QU, CRAG-v2, experimental-project, dice-index, F1-score, weighted shape, contour recognition, object indexing